implement platform agnostic interface and factory
