+++ .gitignore
@@ -162,3 +162,5 @@ cython_debug/
 .*venv*
 /examples/storage
 /.idea
+
+logs/
