+++ examples/adapters_gemini_example.py
@@ -1,26 +1,15 @@
-import json
-import os
-import time
 from jaims import (
-    JAImsAgent,
     JAImsFunctionTool,
-    JAImsDefaultHistoryManager,
     JAImsFunctionToolDescriptor,
     JAImsParamDescriptor,
     JAImsJsonSchemaType,
     JAImsMessage,
+    JAImsDefaultHistoryManager,
 )
 
-from jaims.adapters.openai_adapter import (
-    JAImsOpenaiKWArgs,
-    create_jaims_openai,
-    OpenAITransactionStorageInterface,
+from jaims.adapters.google_generative_ai_adapter import (
+    create_jaims_gemini,
 )
-
-"""
-This example shows how to perform parallel tool calls and how to use the 
-JAImsTransactionStorageInterface to store each transaction with the LLM.
-"""
 
 
 def sum(a: int, b: int):
@@ -47,24 +36,6 @@ def store_multiply(result: int):
     print("----storing multiply----")
     print(result)
     print("-------------------")
-
-
-class FileTransactionStorage(OpenAITransactionStorageInterface):
-
-    def __init__(self, path="storage") -> None:
-        super().__init__()
-        script_dir = os.path.dirname(__file__)
-        self.storage_path = os.path.join(script_dir, path)
-        if not os.path.exists(self.storage_path):
-            os.makedirs(self.storage_path)
-
-    def store_transaction(self, request: dict, response: dict):
-
-        transaction = {"request": request, "response": response}
-        unix_timestamp = str(int(time.time()))
-
-        with open(f"{self.storage_path}/{unix_timestamp}.json", "w") as f:
-            f.write(json.dumps(transaction, indent=4))
 
 
 def main():
@@ -140,20 +111,17 @@ def main():
         ),
     )
 
-    agent = create_jaims_openai(
-        kwargs=JAImsOpenaiKWArgs(
-            model="gpt-4-turbo-2024-04-09",
-            stream=stream,
-        ),
-        transaction_storage=FileTransactionStorage(),
-        history_manager=JAImsDefaultHistoryManager(
-            history=[
-                JAImsMessage.assistant_message(
-                    text="Hello, I am JAIms, your personal assistant."
-                ),
-                JAImsMessage.assistant_message(text="How can I help you today?"),
-            ]
-        ),
+    history_manager = JAImsDefaultHistoryManager(
+        leading_prompts=[
+            JAImsMessage.system_message(
+                "You are JAIms, a helpful assistant that helps the user with math operations."
+            )
+        ],
+    )
+
+    agent = create_jaims_gemini(
+        model="gemini-1.5-pro-latest",
+        history_manager=history_manager,
         tools=[
             sum_func_wrapper,
             multiply_func_wrapper,
