+++ examples/adapters_openai_example.py
@@ -2,26 +2,19 @@ import json
 import os
 import time
 from jaims import (
-    JAImsAgent,
     JAImsFunctionTool,
+    JAImsDefaultHistoryManager,
     JAImsFunctionToolDescriptor,
     JAImsParamDescriptor,
     JAImsJsonSchemaType,
     JAImsMessage,
-    JAImsDefaultHistoryManager,
 )
 
-from jaims.adapters.google_generative_ai_adapter import (
-    JAImsGoogleGenerativeAIAdapter,
-    create_jaims_gemini,
-    GEMINI_1_PRO_LATEST,
-    GEMINI_1_5_FLASH_LATEST,
+from jaims.adapters.openai_adapter import (
+    JAImsOpenaiKWArgs,
+    create_jaims_openai,
+    OpenAITransactionStorageInterface,
 )
-
-"""
-This example shows how to perform parallel tool calls and how to use the 
-JAImsTransactionStorageInterface to store each transaction with the LLM.
-"""
 
 
 def sum(a: int, b: int):
@@ -48,6 +41,24 @@ def store_multiply(result: int):
     print("----storing multiply----")
     print(result)
     print("-------------------")
+
+
+class FileTransactionStorage(OpenAITransactionStorageInterface):
+
+    def __init__(self, path="storage") -> None:
+        super().__init__()
+        script_dir = os.path.dirname(__file__)
+        self.storage_path = os.path.join(script_dir, path)
+        if not os.path.exists(self.storage_path):
+            os.makedirs(self.storage_path)
+
+    def store_transaction(self, request: dict, response: dict):
+
+        transaction = {"request": request, "response": response}
+        unix_timestamp = str(int(time.time()))
+
+        with open(f"{self.storage_path}/{unix_timestamp}.json", "w") as f:
+            f.write(json.dumps(transaction, indent=4))
 
 
 def main():
@@ -123,17 +134,20 @@ def main():
         ),
     )
 
-    history_manager = JAImsDefaultHistoryManager(
-        leading_prompts=[
-            JAImsMessage.system_message(
-                "You are JAIms, a helpful assistant that helps the user with math operations."
-            )
-        ],
-    )
-
-    agent = create_jaims_gemini(
-        model="gemini-1.5-pro-latest",
-        history_manager=history_manager,
+    agent = create_jaims_openai(
+        kwargs=JAImsOpenaiKWArgs(
+            model="gpt-4-turbo-2024-04-09",
+            stream=stream,
+        ),
+        transaction_storage=FileTransactionStorage(),
+        history_manager=JAImsDefaultHistoryManager(
+            history=[
+                JAImsMessage.assistant_message(
+                    text="Hello, I am JAIms, your personal assistant."
+                ),
+                JAImsMessage.assistant_message(text="How can I help you today?"),
+            ]
+        ),
         tools=[
             sum_func_wrapper,
             multiply_func_wrapper,
