+++ jaims/__init__.py
@@ -1,5 +1,7 @@
 from .agent import JAImsAgent
+
 from .default_tool_manager import JAImsDefaultToolManager
+
 from .default_history_manager import (
     JAImsDefaultHistoryManager,
     JAImsLastNHistoryOptimizer,
@@ -16,5 +18,41 @@ from .entities import (
     JAImsParamDescriptor,
     JAImsJsonSchemaType,
     JAImsToolCall,
+    JAImsLLMConfig,
+    JAImsOptions,
+    JAImsImageContentType,
+    JAImsMaxRetriesExceeded,
+    JAImsModelCode,
+    JAImsStreamingMessage,
+    JAImsUnexpectedFunctionCall,
 )
+
 from .interfaces import JAImsToolManager, JAImsHistoryManager, JAImsLLMInterface
+
+__all__ = [
+    "JAImsAgent",
+    "JAImsDefaultToolManager",
+    "JAImsDefaultHistoryManager",
+    "JAImsLastNHistoryOptimizer",
+    "JAImsMessage",
+    "JAImsImageContent",
+    "JAImsContentType",
+    "JAImsMessageRole",
+    "JAImsToolResponse",
+    "JAImsFunctionTool",
+    "JAImsMaxConsecutiveFunctionCallsExceeded",
+    "JAImsFunctionToolDescriptor",
+    "JAImsParamDescriptor",
+    "JAImsJsonSchemaType",
+    "JAImsToolCall",
+    "JAImsLLMConfig",
+    "JAImsOptions",
+    "JAImsImageContentType",
+    "JAImsMaxRetriesExceeded",
+    "JAImsModelCode",
+    "JAImsStreamingMessage",
+    "JAImsUnexpectedFunctionCall",
+    "JAImsToolManager",
+    "JAImsHistoryManager",
+    "JAImsLLMInterface",
+]
