+++ jaims/adapters/__init__.py
@@ -1,1 +1,2 @@
 from .openai_adapter import *
+from .google_generative_ai_adapter import *
