+++ jaims/adapters/google_generative_ai_adapter/__init__.py
@@ -1,6 +1,5 @@
 from .adapter import (
     JAImsGoogleGenerativeAIAdapter,
-    create_jaims_gemini,
 )
 
-from .constants import *
+from .factory import create_jaims_gemini
