+++ jaims/adapters/openai_adapter/__init__.py
@@ -1,8 +1,16 @@
 from .adapter import (
     JAImsOpenaiKWArgs,
-    JAImsOptions,
-    create_jaims_openai,
     JAImsTokenHistoryOptimizer,
     OpenAITransactionStorageInterface,
     JAImsOpenaiAdapter,
 )
+
+from .factory import create_jaims_openai
+
+__all__ = [
+    "JAImsOpenaiKWArgs",
+    "create_jaims_openai",
+    "JAImsTokenHistoryOptimizer",
+    "OpenAITransactionStorageInterface",
+    "JAImsOpenaiAdapter",
+]
