+++ jaims/entities.py
@@ -309,6 +309,105 @@ class JAImsFunctionTool:
         return self.function(**params) if params and self.function else None
 
 
+class JAImsModelCode:
+    """
+    Constants for the LLM names.
+    """
+
+    GEMINI_1_PRO = "gemini-1.0-pro"
+    GEMINI_1_PRO_LATEST = "gemini-1.0-pro-latest"
+    GEMINI_1_PRO_001 = "gemini-1.0-pro-001"
+    GEMINI_1_PRO_VISION = "gemini-1.0-pro-vision"
+    GEMINI_1_PRO_VISION_LATEST = "gemini-1.0-pro-vision-latest"
+    GEMINI_1_5_FLASH_LATEST = "gemini-1.5-flash-latest"
+    GEMINI_1_5_PRO = "gemini-1.5-pro"
+    GEMINI_1_5_PRO_LATEST = "gemini-1.5-pro-latest"
+    GPT_3_5_TURBO = "gpt-3.5-turbo"
+    GPT_4 = "gpt-4"
+    GPT_4_32K = "gpt-4-32k"
+    GPT_4_0613 = "gpt-4-0613"
+    GPT_4_32K_0613 = "gpt-4-32k-0613"
+    GPT_4_TURBO = "gpt-4-turbo"
+    GPT_4_TURBO_2024_04_09 = "gpt-4-turbo-2024-04-09"
+    GPT_4_TURBO_PREVIEW = "gpt-4-turbo-preview"
+    GPT_4_0125_PREVIEW = "gpt-4-0125-preview"
+    GPT_4_1106_PREVIEW = "gpt-4-1106-preview"
+    GPT_4_VISION_PREVIEW = "gpt-4-vision-preview"
+    GPT_4_1106_VISION_PREVIEW = "gpt-4-1106-vision-preview"
+    GPT_4o = "gpt-4o"
+    GPT_4o_2024_05_13 = "gpt-4o-2024-05-13"
+
+
+class JAImsLLMConfig:
+    def __init__(
+        self,
+        temperature: float = 0.5,
+        max_tokens: int = 1024,
+        response_format: Optional[Dict[str, Any]] = None,
+    ):
+        self.temperature = temperature
+        self.max_tokens = max_tokens
+        self.response_format = response_format
+
+
+class JAImsOptions:
+    """
+    Config options for the JAImsAgent when calling the remote LLM.
+    Exponential backoff is calculated using the formula: min(delay * exponential_base, exponential_cap) * (1 + jitter * random())
+
+    Args:
+        max_retries (int): The maximum number of retries after a failing a call.
+        retry_delay (int): The delay between each retry in case of failure without exponential backoff.
+        exponential_base (int): The base for exponential backoff calculation.
+        exponential_delay (int): The initial delay, in seconds, to multiply by the base for exponential backoff.
+        exponential_cap (Optional[int]): The maximum value, in seconds, for exponential backoff delay. Leave None to let it grow indefinitely.
+        jitter (bool): Whether to add a small jitter to the delay (to avoid concurrent firing), in case of exponential backoff, in the worst case, it will be 2x the delay.
+    """
+
+    def __init__(
+        self,
+        max_retries=15,
+        retry_delay=10,
+        exponential_base: int = 2,
+        exponential_delay: int = 1,
+        exponential_cap: Optional[int] = None,
+        jitter: bool = True,
+    ):
+        self.max_retries = max_retries
+        self.retry_delay = retry_delay
+        self.exponential_base = exponential_base
+        self.exponential_delay = exponential_delay
+        self.exponential_cap = exponential_cap
+        self.jitter = jitter
+
+    def copy_with_overrides(
+        self,
+        max_retries: Optional[int] = None,
+        retry_delay: Optional[int] = None,
+        exponential_base: Optional[int] = None,
+        exponential_delay: Optional[int] = None,
+        exponential_cap: Optional[int] = None,
+        jitter: Optional[bool] = None,
+    ) -> JAImsOptions:
+        """
+        Returns a new JAImsOptions instance with the passed parameters overridden.
+        """
+        return JAImsOptions(
+            max_retries=max_retries if max_retries else self.max_retries,
+            retry_delay=retry_delay if retry_delay else self.retry_delay,
+            exponential_base=(
+                exponential_base if exponential_base else self.exponential_base
+            ),
+            exponential_delay=(
+                exponential_delay if exponential_delay else self.exponential_delay
+            ),
+            exponential_cap=(
+                exponential_cap if exponential_cap else self.exponential_cap
+            ),
+            jitter=jitter if jitter else self.jitter,
+        )
+
+
 # ----------
 # exceptions
 # ----------
@@ -337,4 +436,19 @@ class JAImsUnexpectedFunctionCall(Exception):
 
     def __init__(self, func_name):
         message = f"Unexpected function call: {func_name}"
+        super().__init__(message)
+
+
+class JAImsMaxRetriesExceeded(Exception):
+    """
+    Exception raised when the maximum number of retries is performed by an adapter client.
+
+    Attributes:
+        max_consecutive_calls -- maximum number of consecutive calls allowed
+    """
+
+    def __init__(self, max_retries, last_exception=None):
+        message = (
+            f"Max retries exceeded: {max_retries}, last exception: {last_exception}"
+        )
         super().__init__(message)
